Enhancements to route generator (#27)
This pull request introduces several enhancements to the route
generation process, focusing on making routes type-safe. It also
includes improvements to support query parameters, automatic suspense
wrapping, and fixes for Hot Module Replacement (HMR).

## Changes
1. **Type-Safe Routes**:
- For each route, an accompanying file is generated into the `@`
directory that includes type information.
- The generated file is intended to be imported and used in the route
file.
- The `RouteGenerator` ensures that route files import the correct
accompanying file.

2. **Support for Query Parameters**:
   - Query parameters are type-safe.

3. **Automatic Suspense Wrapping**:
   - Routes are now automatically wrapped in `Suspense`.
- The suspense fallback can be specified by exporting a `LoadingState`
component from the route file.
- To opt out of suspense wrapping, export `const LoadingState = false`.

4. **Route Generation CLI**:
- The route generation CLI script has been moved to the `scripts`
directory.
   - The CLI now supports watch mode.

5. **Fixes for Hot Module Replacement (HMR)**:
   - The `getData` function is no longer exported from the route files.
   - The previously unnamed default export is now named `Route`.
- To avoid exporting the `getData` function, the `RouteGenerator` now
copies a modified version of the route file to the `@` directory in
which the `getData` function is exported. The route config file then
uses this file instead of the original to access the `getData` function.

Note that the generated files are not checked in to version control
anymore. As a consequence, you will need to build the routes before the
app can be run. The appropriate commands to do this have been added to
`package.json`.